[virt_agent] 统一修复安全问题及优化UT用例

+++ src/addons/virt_agent/common/message/include/ham_migrate_dst_info_message.h
@@ -44,7 +44,7 @@ public:
         return hamMigrateDstInfo;
     }
 
-    inline void SetHamMigrateDstInfo(int dstPid, const std::string& dstNodeId)
+    inline void SetHamMigrateDstInfo(uint64_t dstPid, const std::string& dstNodeId)
     {
         hamMigrateDstInfo.dstPid = dstPid;
         hamMigrateDstInfo.dstNodeId = dstNodeId;
